backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.4.2
python-dotenv==1.0.0
sqlalchemy==2.0.23
chromadb==0.4.18
transformers==4.35.2
torch==2.1.0
sentence-transformers==2.2.2
python-multipart==0.0.6
requests==2.31.0